+++ Extensions/RecluseEdit.Extensions.DeepSeek/Views/DeepSeekChatView.xaml.cs
@@ -317,3 +317,4 @@ public partial class DeepSeekChatView : UserControl
 
     #endregion
 }
+